requirements.txt
# Web Automation
selenium>=4.10.0
webdriver-manager>=4.0.0

# UI
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0